9/1/2022 2

+++ database/migrations/2021_12_27_201701_create_client_table.php
@@ -19,7 +19,7 @@ class CreateClientTable extends Migration
             $table->string('username')->unique();
             $table->text('password');
             $table->text('name');
-            $table->integer('currentCredit')->default(0);
+            $table->float('currentCredit', 8, 2)->default(0);
             $table->text('token');
 
             $table->timestamps();
